Fix Download Sponsorship Prospectus button view
We had a text-center class on the button but we didn't had the required
css properties for that class, also while centring the button we were
giving left and right margin too but that makes the button text to move
to the next line(in the mobile view). This patch updates the styles
by simply adding the top and bottom margin and to center align it just
uses the text-align property and It also removes the unnecessary style
attribute.